Added `Mutation`s to edit `Document`.
